+++ examples/react/todo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # examples/react/todo
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [[`5cafaf4`](https://github.com/TanStack/db/commit/5cafaf4f73018599ce799778029833d3fa83dbc9)]:
+  - @tanstack/db-collections@0.0.15
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [[`f6abe9b`](https://github.com/TanStack/db/commit/f6abe9b94b890487fe960bd72a89e4a75de89d46)]:
+  - @tanstack/db-collections@0.0.14
+  - @tanstack/react-db@0.0.12
+
 ## 0.0.15
 
 ### Patch Changes